Конец разработки

+++ app/controllers/plan_events_controller.rb
@@ -1,9 +1,9 @@
 # app/controllers/plans/plan_events_controller.rb
   class PlanEventsController < ApplicationController
+    before_action :authenticate_user!
     before_action :set_plan , except: [:destroy]
     before_action :set_plan_event, only: [:destroy]
     before_action :store_referer, only:[:destroy]
-    before_action :check_permissions
     
     def create
       event_department = ApprovedEventDepartment.find(params[:event_department_id])
@@ -28,11 +28,14 @@
     
     def destroy
       event_title = @plan_event.event_department.event.title
-      
-      if @plan_event.destroy
-        redirect_to stored_referer, notice: "Мероприятие '#{event_title}' удалено из плана"
+      if can?(:delete, @plan_event)
+        if @plan_event.destroy
+          redirect_to stored_referer, notice: "Мероприятие '#{event_title}' удалено из плана"
+        else
+          redirect_to stored_referer, alert: "Не удалось удалить мероприятие из плана"
+        end
       else
-        redirect_to stored_referer, alert: "Не удалось удалить мероприятие из плана"
+        
       end
     end
 
@@ -95,19 +98,6 @@
 
     private
 
-    def can_access_department_plans?(user, department)
-      # Проверяем права на просмотр планов подразделения
-      # 1. Пользователь должен иметь роль в этом подразделении ИЛИ
-      # 2. Иметь право на управление планами дочерних подразделений
-      
-      # Проверка по ролям пользователя
-      user_has_access = user.departments.include?(department)
-      
-      # ИЛИ проверка прав ABAC
-      abac_access = can?(:view, Plan.new(department: department))
-      
-      user_has_access || abac_access
-    end
 
     def format_plan_for_json(plan, coverage_type = :full_coverage)
       {
@@ -123,17 +113,6 @@
         coverage_badge: coverage_badge(coverage_type)
       }
     end
-
-    def coverage_badge(coverage_type)
-      case coverage_type
-      when :full_coverage
-        { text: "Полное покрытие", class: "green", icon: "check circle" }
-      when :partial_coverage
-        { text: "Частичное покрытие", class: "yellow", icon: "warning circle" }
-      else
-        { text: "Неизвестно", class: "grey", icon: "question circle" }
-      end
-    end
     
     def set_plan
       @plan = Plan.find(params[:plan_id])
@@ -142,14 +121,5 @@
     def set_plan_event
       @plan_event = PlanEvent.find(params[:id])
       @plan = @plan_event.plan
-    end
-    
-    def check_permissions
-      case action_name.to_sym
-      when :create
-        authorize_action(:add_event, @plan.department)
-      when :destroy
-        authorize_action(:remove_event, @plan.department)
-      end
     end
   end
